Improve voice call and audio recording quality
- improved call quality (speakerphone/earpiece/microphones)
- improved microphone during video and sound recording
- improved speaker

Value 93 for DEC1 and DEC2 is used during video and sound recording
Value 90 for DEC1 and DEC2 is used during the call
Value 89 for DEC1 and DEC2 is used is used while speakerphone is turned on.

Microphone important path names:
handset-mic is used for sound recording.
handset-stereo-dmic-ef is used for video recording
voice-dmic-ef is used for voice call
voice-speaker-dmic-ef is used for voice call while speakerphone is turned on